refactor: Move stamp module into state package
