+++ docs/api/state.md
@@ -3,3 +3,5 @@
 ::: napt.state.cache
 
 ::: napt.state.deployment
+
+::: napt.state.stamp
