+++ docs/api/upload.md
@@ -3,5 +3,3 @@
 ::: napt.upload.manager
 
 ::: napt.upload.intunewin
-
-::: napt.upload.stamp
