+++ napt/promote/applier.py
@@ -76,7 +76,7 @@ from napt.state import (
     load_deployment_state,
     save_deployment_state,
 )
-from napt.upload.stamp import ENTRY_INSTALL, ENTRY_UPDATE, find_stamped_app
+from napt.state.stamp import ENTRY_INSTALL, ENTRY_UPDATE, find_stamped_app
 
 __all__ = ["apply_plan", "load_plan_file"]
