+++ napt/state/stamp.py
